Create documentation + refactoring
