+++ library/src/main/java/com/wultra/android/mtokensdk/oidc/OidcService.kt
@@ -16,31 +16,26 @@ import com.google.gson.GsonBuilder
 import com.wultra.android.mtokensdk.api.oidc.ConfigResponse
 import com.wultra.android.mtokensdk.api.oidc.OidcApi
 import com.wultra.android.mtokensdk.api.oidc.model.OidcConfigRequest
-import com.wultra.android.mtokensdk.api.oidc.model.OidcConfigResponse
 import com.wultra.android.mtokensdk.log.WMTLogger
-import com.wultra.android.mtokensdk.oidc.models.OidcPowerAuthActivationAttributes
-import com.wultra.android.mtokensdk.oidc.models.OidcAuthorizationRequest
-import com.wultra.android.mtokensdk.oidc.utils.PKCEUtils
-import com.wultra.android.mtokensdk.oidc.utils.RandomGeneratorUtils
-import com.wultra.android.mtokensdk.oidc.utils.OidcUriUtils
+import com.wultra.android.mtokensdk.oidc.models.OidcAuthorizationData
+import com.wultra.android.mtokensdk.oidc.models.OidcConfig
+import com.wultra.android.mtokensdk.oidc.models.PKCECodes
+import com.wultra.android.mtokensdk.oidc.utils.OidcUtils
+import com.wultra.android.mtokensdk.oidc.utils.toOidcConfig
 import com.wultra.android.powerauth.networking.IApiCallResponseListener
 import com.wultra.android.powerauth.networking.OkHttpBuilderInterceptor
 import com.wultra.android.powerauth.networking.UserAgent
 import com.wultra.android.powerauth.networking.error.ApiError
 import com.wultra.android.powerauth.networking.error.ApiErrorException
 import com.wultra.android.powerauth.networking.tokens.IPowerAuthTokenProvider
-import io.getlime.security.powerauth.exception.PowerAuthMissingConfigException
-import io.getlime.security.powerauth.networking.interfaces.ICancelable
-import io.getlime.security.powerauth.networking.response.ICreateActivationListener
-import io.getlime.security.powerauth.sdk.PowerAuthActivation
 import io.getlime.security.powerauth.sdk.PowerAuthSDK
 import okhttp3.OkHttpClient
 
 class OidcService(
+    powerAuthSDK: PowerAuthSDK,
+    appContext: Context,
     okHttpClient: OkHttpClient,
     baseURL: String,
-    appContext: Context,
-    powerAuthSDK: PowerAuthSDK,
     tokenProvider: IPowerAuthTokenProvider? = null,
     userAgent: UserAgent? = null,
     gsonBuilder: GsonBuilder?
@@ -75,53 +70,6 @@ class OidcService(
         }
 
     /**
-     * Prepares the OIDC (OpenID Connect) activation process by performing the following steps:
-     *
-     * 1. Retrieves the OIDC configuration using the provided `providerId`.
-     * 2. Generates PKCE (Proof Key for Code Exchange) codes if PKCE is enabled for the configuration.
-     * 3. Creates a `nonce` (a unique value to mitigate replay attacks) and `state` (to maintain state between the request and callback).
-     * 4. Constructs an authorization URI, which is used to generate a URL that should be opened in the browser for user authentication.
-     *
-     * Upon successful execution, the function invokes the provided callback with an [OidcAuthorizationRequest] object containing the authorization URI
-     * and other details required for the OIDC flow. If any step fails, the callback is invoked with a [Result.failure].
-     *
-     * @param providerId A unique identifier for the configuration record, used as a key to retrieve the configuration.
-     * @param callback A callback that receives the result of the OIDC preparation. On success, it returns [OidcAuthorizationRequest].
-     *                 On failure, it returns an appropriate [Throwable].
-     */
-    fun prepareOidcActivation(providerId: String, callback: (Result<OidcAuthorizationRequest>) -> Unit) {
-        // Get configuration based on providerId
-        getConfig(providerId) { configResult ->
-            configResult.onSuccess { oidcConfig ->
-
-                // If pkceEnabled create PKCE Codes or continue with null
-                val resultPKCE = PKCEUtils.create(oidcConfig.pkceEnabled, 32)
-                resultPKCE.onSuccess { pkceCodes ->
-                    val nonce = RandomGeneratorUtils.getRandomBase64UrlSafe(32)
-                    val state = RandomGeneratorUtils.getRandomBase64UrlSafe(32)
-
-                    // Finally create authorizeUri
-                    val authorizeUri = OidcUriUtils.createAuthorizationUri(oidcConfig, nonce, state, pkceCodes)
-                    authorizeUri.onSuccess { authUri ->
-
-                        // Callback with success
-                        callback(Result.success(OidcAuthorizationRequest(authUri, providerId, nonce, state, pkceCodes?.codeVerifier)))
-                    }.onFailure { error ->
-                        WMTLogger.e("OIDC: Failed to create authorization Uri: ${error.message}")
-                        callback(Result.failure(error))
-                    }
-                }.onFailure { error ->
-                    WMTLogger.e("OIDC: Failed to create PKCE codes: ${error.message}")
-                    callback(Result.failure(error))
-                }
-            }.onFailure { error ->
-                WMTLogger.e("OIDC: Failed to get config: ${error.message}")
-                callback(Result.failure(error))
-            }
-        }
-    }
-
-    /**
      * Get configuration of OIDC.
      *
      * @param providerId Predefined identification of the configuration record, used as a key for the configuration
@@ -129,44 +77,65 @@ class OidcService(
      */
     fun getConfig(
         providerId: String,
-        callback: (Result<OidcConfigResponse>) -> Unit
+        callback: (Result<OidcConfig>) -> Unit
     ) {
         oidcApi.getConfig(
             OidcConfigRequest(providerId),
             object : IApiCallResponseListener<ConfigResponse> {
                 override fun onFailure(error: ApiError) {
+                    WMTLogger.e("OIDC: Failed to get config: ${error.e.message}")
                     callback(Result.failure(ApiErrorException(error)))
                 }
 
                 override fun onSuccess(result: ConfigResponse) {
-                    callback(Result.success(result.responseObject))
+                    callback(Result.success(result.responseObject.toOidcConfig()))
                 }
             }
         )
     }
 
     /**
-     * Create a new activation by calling a PowerAuth Standard RESTful API.
+     * Prepares the OIDC (OpenID Connect) authorization data required for the activation process.
      *
-     * @receiver PowerAuthSDK
-     * @param attributes Data object containing the information required for the activation creation.
-     *  - to create attributes see [OidcUriUtils.processDeeplinkOidc]
-     * @param listener A callback listener called when the process finishes - it contains an activation fingerprint in case of success or an error in case of failure.
-     * @return ICancelable object associated with the running HTTP request.
-     * @throws PowerAuthMissingConfigException – thrown in case configuration is not present.
+     * The function performs the following steps:
+     * 1. Generates PKCE (Proof Key for Code Exchange) codes if PKCE is enabled in the provided configuration.
+     * 2. Creates a `nonce` (a unique value to mitigate replay attacks) and a `state` (to maintain state between the request and callback).
+     * 3. Creates data containing an authorization URI that shall be used to open a browser for user authentication.
+     *
+     * @param oidcConfig The OIDC configuration, either retrieved dynamically (e.g., using `getConfig`) or defined statically.
+     * @param callback A callback that receives the result of the authorization data preparation.
+     *                 - On success: Returns [OidcAuthorizationData]
+     *                 - On failure: Returns an appropriate [Throwable].
      */
-    @Throws(PowerAuthMissingConfigException::class)
-    fun PowerAuthSDK.createOidcActivation(attributes: OidcPowerAuthActivationAttributes, listener: ICreateActivationListener): ICancelable? {
-        val activation =
-            PowerAuthActivation
-                .Builder
-                .oidcActivation(
-                    attributes.providerId,
-                    attributes.code,
-                    attributes.nonce,
-                    attributes.codeVerifier
-                )
-                .build()
-        return createActivation(activation, listener)
+    fun prepareOidcAuthorizationData(oidcConfig: OidcConfig, callback: (Result<OidcAuthorizationData>) -> Unit) {
+        // If pkceEnabled create PKCE Codes or continue with null
+        val resultPKCE = createPKCE(oidcConfig.pkceEnabled, 32)
+        resultPKCE.onSuccess { pkceCodes ->
+            val nonce = OidcUtils.getRandomBase64UrlSafe(32)
+            val state = OidcUtils.getRandomBase64UrlSafe(32)
+
+            // Finally create authorizeUri
+            val authorizeUri = OidcUtils.createAuthorizationUri(oidcConfig, nonce, state, pkceCodes)
+            authorizeUri.onSuccess { authUri ->
+
+                // Callback with success
+                callback(Result.success(OidcAuthorizationData(authUri, oidcConfig.providerId, nonce, state, pkceCodes?.codeVerifier)))
+            }.onFailure { error ->
+                WMTLogger.e("OIDC: Failed to create authorization Uri: ${error.message}")
+                callback(Result.failure(error))
+            }
+        }.onFailure { error ->
+            WMTLogger.e("OIDC: Failed to create PKCE codes: ${error.message}")
+            callback(Result.failure(error))
+        }
+    }
+
+    // Helper method to return Result null if PKCE is not enabled
+    private fun createPKCE(enabled: Boolean, dataLength: Int): Result<PKCECodes?> {
+        return if (!enabled) {
+            Result.success(null)
+        } else {
+            OidcUtils.createPKCE(dataLength)
+        }
     }
 }
